release: 0.3.2
In the previous release, the version was directly read from `pyproject.toml`, with the help of the `tomli` package. This led to filesystem errors and could not be used. Now, however, the version is specified in `__init__.py` and the version is bumped accordingly using the `poetry version [major|minor|patch]` command. Also, with the help of a poetry plugin called `poetry-bumpversion` (https://github.com/monim67/poetry-bumpversion), it is now possible to dynamically update both the `__init__.py` version and `pyproject.toml` version at the same time.